Split and rename default package in OTEL
