+++ cmd/opentelemetry/app/defaultconfig/default_config.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package defaults
+package defaultconfig
 
 import (
 	"fmt"
