+++ cmd/opentelemetry/app/defaultconfig/merge.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package defaults
+package defaultconfig
 
 import (
 	"github.com/imdario/mergo"
