+++ cmd/opentelemetry/app/defaultconfig/testdata/addprocessor-merged.yaml
