+++ cmd/opentelemetry/app/defaultconfig/testdata/addprocessor.yaml
