+++ cmd/opentelemetry/app/defaultconfig/testdata/emptyoverride-merged.yaml
