+++ cmd/opentelemetry/app/defaultconfig/testdata/emptyoverride.yaml
