+++ cmd/opentelemetry/app/defaultconfig/testdata/multiplecomponents-merged.yaml
