+++ cmd/opentelemetry/app/defaultconfig/testdata/multiplecomponents-override.yaml
