+++ cmd/opentelemetry/app/defaultconfig/testdata/multiplecomponents.yaml
