+++ cmd/opentelemetry/cmd/ingester/main.go
@@ -27,7 +27,8 @@ import (
 
 	jflags "github.com/jaegertracing/jaeger/cmd/flags"
 	"github.com/jaegertracing/jaeger/cmd/opentelemetry/app"
-	"github.com/jaegertracing/jaeger/cmd/opentelemetry/app/defaults"
+	"github.com/jaegertracing/jaeger/cmd/opentelemetry/app/defaultcomponents"
+	"github.com/jaegertracing/jaeger/cmd/opentelemetry/app/defaultconfig"
 	"github.com/jaegertracing/jaeger/cmd/opentelemetry/app/receiver/kafkareceiver"
 	jconfig "github.com/jaegertracing/jaeger/pkg/config"
 	"github.com/jaegertracing/jaeger/pkg/version"
@@ -56,10 +57,10 @@ func main() {
 		storageType = "cassandra"
 	}
 
-	cmpts := defaults.Components(v)
+	cmpts := defaultcomponents.Components(v)
 	cfgFactory := func(otelViper *viper.Viper, f config.Factories) (*configmodels.Config, error) {
-		cfgConfig := defaults.ComponentSettings{
-			ComponentType: defaults.Ingester,
+		cfgConfig := defaultconfig.ComponentSettings{
+			ComponentType: defaultconfig.Ingester,
 			Factories:     cmpts,
 			StorageType:   storageType,
 		}
@@ -72,7 +73,7 @@ func main() {
 			if err != nil {
 				return nil, err
 			}
-			err = defaults.MergeConfigs(cfg, otelCfg)
+			err = defaultconfig.MergeConfigs(cfg, otelCfg)
 			if err != nil {
 				return nil, err
 			}
